# openapi.yaml
openapi: 3.0.0
info:
  title: whatsapp-multi-gateway API
  description: |
    Production-grade multi-tenant WhatsApp API Gateway + Web Dashboard built on
    Baileys v7. Provides REST endpoints for managing WhatsApp instances, sending
    and receiving messages, HMAC-signed webhooks, real-time Socket.IO events,
    authentication, RBAC and audit logging.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3333
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/instances:
    post:
      operationId: createInstance
      summary: Create a WhatsApp instance
      description: |
        Creates and starts a new WhatsApp connection. When `usePairingCode` is
        true and `phoneNumber` is provided a pairing code is returned instead of
        a QR code.
      tags: [Instances]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceRequest'
            example:
              name: sales-line
              webhookUrl: https://example.com/webhook
              webhookSecret: some-secret
              usePairingCode: false
      responses:
        '201':
          description: Instance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  instance:
                    id: inst_4f9d2c1b
                    instanceId: inst_4f9d2c1b
                    name: sales-line
                    status: qr_ready
                    phoneNumber: null
                    pushName: null
                    platform: null
                    connectionState: connecting
                    webhookUrl: https://example.com/webhook
                    createdAt: '2026-01-01T12:00:00.000Z'
                  qr: '2@...'
                requestId: req_4f9d2c1b
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner, admin or operator)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    get:
      operationId: listInstances
      summary: List instances for the tenant
      tags: [Instances]
      responses:
        '200':
          description: List of instances
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  - id: inst_4f9d2c1b
                    instanceId: inst_4f9d2c1b
                    name: sales-line
                    status: connected
                    phoneNumber: '15551234567'
                    pushName: Sales Line
                    webhookUrl: https://example.com/webhook
                    connectionState: open
                requestId: req_4f9d2c1b
  /api/v1/instances/{id}:
    get:
      operationId: getInstance
      summary: Get a single instance
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  instance:
                    id: inst_4f9d2c1b
                    instanceId: inst_4f9d2c1b
                    name: sales-line
                    status: connected
                    phoneNumber: '15551234567'
                    connectionState: open
                    webhookUrl: https://example.com/webhook
                    lastConnectedAt: '2026-01-01T12:00:00.000Z'
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error:
                  code: NOT_FOUND
                  message: "Instance with id 'inst_xxx' not found"
                requestId: req_4f9d2c1b
    delete:
      operationId: deleteInstance
      summary: Delete an instance
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  deleted: true
                  instanceId: inst_4f9d2c1b
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner or admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/qr:
    get:
      operationId: getInstanceQr
      summary: Get the current QR code for an instance
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current QR code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  qr: '2@1x9c...'
                  instanceId: inst_4f9d2c1b
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: No QR code currently available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                success: false
                error:
                  code: CONFLICT
                  message: No QR code is currently available for this instance
                requestId: req_4f9d2c1b
  /api/v1/instances/{id}/qr/status:
    get:
      operationId: getInstanceQrStatus
      summary: Get instance status and connected flag
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  instanceId: inst_4f9d2c1b
                  status: connected
                  connected: true
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/pairing:
    post:
      operationId: requestPairingCode
      summary: Request a WhatsApp pairing code
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairingRequest'
            example:
              phoneNumber: '15551234567'
      responses:
        '200':
          description: Pairing code issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  pairingCode: ABCDEFGH
                  phoneNumber: '15551234567'
                requestId: req_4f9d2c1b
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner, admin or operator)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/disconnect:
    post:
      operationId: disconnectInstance
      summary: Disconnect an instance
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance disconnected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  disconnected: true
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner, admin or operator)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/reconnect:
    post:
      operationId: reconnectInstance
      summary: Reconnect an instance
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance reconnected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  reconnected: true
                  instanceId: inst_4f9d2c1b
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner, admin or operator)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/logout:
    post:
      operationId: logoutInstance
      summary: Log out of WhatsApp and delete session credentials
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  loggedOut: true
                  instanceId: inst_4f9d2c1b
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner or admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/chats:
    get:
      operationId: listChats
      summary: List chats for an instance
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of chats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  chats: []
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/groups:
    get:
      operationId: listGroups
      summary: List groups for an instance
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  groups: []
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/webhook:
    get:
      operationId: getInstanceWebhook
      summary: Read webhook configuration
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  webhookUrl: https://example.com/webhook
                  hasSecret: true
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Instance not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    put:
      operationId: updateInstanceWebhook
      summary: Update webhook URL and/or secret
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookConfigRequest'
            example:
              webhookUrl: https://example.com/webhook
              webhookSecret: new-secret
      responses:
        '200':
          description: Webhook configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  updated: true
                requestId: req_4f9d2c1b
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner, admin or operator)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/instances/{id}/webhook/test:
    post:
      operationId: testInstanceWebhook
      summary: Send a test webhook delivery
      tags: [Instances]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test webhook dispatched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
              example:
                success: true
                data:
                  tested: true
                  message: test webhook dispatched
                requestId: req_4f9d2c1b
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden (role must be owner, admin or operator)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer JWT obtained from POST /api/v1/auth/login. Any protected endpoint
        also accepts an X-API-Key header instead of a Bearer token.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        API key created via POST /api/v1/auth/api-keys. The plaintext key starts
        with 'wag_' and is shown once at creation. Keys are hashed at rest.
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          description: Response payload (shape varies per endpoint)
        requestId:
          type: string
          example: req_4f9d2c1b
      required: [success, data, requestId]
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          properties:
            code:
              type: string
              example: UNAUTHORIZED
            message:
              type: string
              example: Unauthorized
            details:
              type: object
              additionalProperties: true
          required: [code, message]
        requestId:
          type: string
          example: req_4f9d2c1b
      required: [success, error, requestId]
...
    CreateInstanceRequest:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: sales-line
        webhookUrl:
          type: string
          format: uri
          example: https://example.com/webhook
        webhookSecret:
          type: string
          minLength: 8
          example: some-secret
        usePairingCode:
          type: boolean
          default: false
        phoneNumber:
          type: string
          example: '15551234567'
...
    WebhookConfigRequest:
      type: object
      required: [webhookUrl]
      additionalProperties: false
      properties:
        webhookUrl:
          type: string
          format: uri
          example: https://example.com/webhook
        webhookSecret:
          type: string
          minLength: 8
          example: some-secret
    PairingRequest:
      type: object
      required: [phoneNumber]
      additionalProperties: false
      properties:
        phoneNumber:
          type: string
          minLength: 1
          example: '15551234567'
...